Batch gets own mod, adding multiplicity
