+++ traversal/executor/mod.rs
@@ -16,6 +16,7 @@ mod function_executor;
 mod instruction;
 pub mod pattern_executor;
 pub mod program_executor;
+mod batch;
 
 #[derive(Debug, Copy, Clone, Eq, PartialEq, Hash, Ord, PartialOrd)]
 pub struct Position {
